Abstract config logic into prepareConfig()
This was essentially the same
in 2 places, which was getting
hard to maintain.
