+++ src/utils/prepareConfig.ts
@@ -1,0 +1,56 @@
+import * as fs from 'fs';
+import * as path from 'path';
+import * as vscode from 'vscode';
+import * as yaml from 'js-yaml';
+import TelemetryReporter from 'vscode-extension-telemetry';
+import getConfigFromPath from './getConfigFromPath';
+import getDynamicConfigFilePath from './getDynamicConfigFilePath';
+import getProcessedConfig from './getProcessedConfig';
+import getProcessFilePath from './getProcessFilePath';
+import replaceDynamicConfigOrb from './replaceDynamicConfigOrb';
+import writeProcessFile from './writeProcessFile';
+
+export default function prepareConfig(
+  configFilePath: string,
+  reporter: TelemetryReporter,
+  suppressMessage?: boolean
+): { processedConfig: string; processError: string } {
+  let processedConfig = '';
+  let processError = '';
+
+  try {
+    const processFilePath = getProcessFilePath(configFilePath);
+
+    if (!fs.existsSync(path.dirname(processFilePath))) {
+      fs.mkdirSync(path.dirname(processFilePath), { recursive: true });
+    }
+    fs.writeFileSync(
+      processFilePath,
+      yaml.dump(replaceDynamicConfigOrb(getConfigFromPath(configFilePath)))
+    );
+
+    processedConfig = getProcessedConfig(processFilePath);
+    writeProcessFile(processedConfig, processFilePath);
+
+    const dynamicConfigFilePath = getDynamicConfigFilePath(configFilePath);
+    if (fs.existsSync(dynamicConfigFilePath)) {
+      writeProcessFile(
+        getProcessedConfig(dynamicConfigFilePath),
+        dynamicConfigFilePath
+      );
+    }
+  } catch (e) {
+    processError = (e as ErrorWithMessage)?.message;
+    if (!suppressMessage) {
+      vscode.window.showErrorMessage(
+        `There was an error processing the CircleCI config: ${
+          (e as ErrorWithMessage)?.message
+        }`
+      );
+    }
+
+    reporter.sendTelemetryErrorEvent('writeProcessFile');
+  }
+
+  return { processedConfig, processError };
+}
